fix(Metadata): updates all styles
